+++ packages/flutter_scene_layout3d/lib/testing.dart
@@ -30,6 +30,7 @@ library;
 
 export 'src/testing/finders.dart'
     show CommonLayout3dFinders, Layout3dFinder, find3d;
+export 'src/testing/geometry.dart' show Layout3dTestGeometry;
 export 'src/testing/matchers.dart'
     show hasSize3d, hasSizeDp, isReachable3d, standsOnItsPanel3d;
 export 'src/testing/tester.dart' show Layout3dWidgetTester, cameraFacing3d;